Don't rely on onexpire callback to re-take lock
The callback is fired by a timer in the client process, it's not
guaranteed to fire before the lock expires on the server.
Instead, renew the lock at a sensible proportion of the lock timeout.
There's no reason to wait until the very last second (we're talking
about a very small amount of traffic).